+++ shared/lib_pv_io_manager.cpp
@@ -696,7 +696,7 @@ PVSystem_IO::PVSystem_IO(compute_module* cm, std::string cmName, Simulation_IO *
 		for (size_t n_subarray = 0; n_subarray < Subarrays.size(); n_subarray++) //jmf update this so that all subarray markers are consistent, get rid of "enable" check
 			if (Subarrays[n_subarray]->enable)
 				if (Subarrays[n_subarray]->mpptInput == (int)mppt)
-					mppt_n.push_back(n_subarray);
+					mppt_n.push_back((int)n_subarray);
 		if (mppt_n.size() < 1)
 			throw compute_module::exec_error(cmName, "At least one subarray must be assigned to each inverter MPPT input.");
 		mpptMapping.push_back(mppt_n); //add the subarrays on this input to the total mppt mapping vector
